temporal attention fix
A typo in the original code meant that the value tensors for the temporal attention step were identical to the input instead of being multiplied by a learned projection matrix (v = x rather than v = Wx). The original code is kept to facilitate replication and can be used by setting use_original_code=True, but is not recommended.